Remove multiple blog theme files
The changes in the code indicate that numerous blog theme files were deleted. The purpose of this action is unclear without context or comments. If it was intended, the blog site will likely see major modifications, especially if these were main functionalities. These changes should be tested extensively to ensure the site still runs as expected.